Refactor tests
to use regions fixture
