+++ fixtures/regions.js
@@ -1,0 +1,8 @@
+const regions = [
+  { id: 1, name: '서울' },
+  { id: 2, name: '부산' },
+  { id: 3, name: '대구' },
+  { id: 4, name: '대전' },
+];
+
+export default regions;
