+++ src/Regions.test.jsx
@@ -4,14 +4,9 @@ import { render } from '@testing-library/react';
 
 import Regions from './Regions';
 
-test('Regions', () => {
-  const regions = [
-    { id: 1, name: '서울' },
-    { id: 2, name: '부산' },
-    { id: 3, name: '대구' },
-    { id: 4, name: '대전' },
-  ];
+import regions from '../fixtures/regions';
 
+test('Regions', () => {
   const { queryByText } = render(<Regions regions={regions} />);
 
   expect(queryByText(/서울/)).not.toBeNull();
